WIP FT_EXPORT:
WIP export, working on adding an already existing var.
Also thinking of a way to implement easily += for haters.

Changes to be committed:
	modified:   ../../Makefile
	modified:   ../../includes/minishell.h
	modified:   main_test_staf.c
	new file:   ms_builtin_utils.c
	new file:   ms_export.c
	modified:   ms_unset.c
	modified:   ../env/ms_env_to_list.c
