+++ Makefile
@@ -88,10 +88,12 @@ teststaf:			$(LIBFT)
 					@clear
 					$(CC) -Wall -Wextra -g3 \
 					./sources/builtins/ms_unset.c	\
+					./sources/builtins/ms_export.c	\
 					./sources/env/ms_env_to_list.c  \
 					./sources/builtins/main_test_staf.c \
 					./sources/env/ms_free_env.c \
 					./sources/env/ms_print_env.c \
+					./sources/builtins/ms_builtin_utils.c \
 					-I ./includes -I ./libft/includes libft/libft.a
 					valgrind --leak-check=full ./a.out
